requirements.txt
flask==2.0.1
line-bot-sdk==3.11.0
mysql-connector-python==8.0.33